# Improve performance of `syntaxKindsByLine`