--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -16,7 +16,7 @@
   [Daniel Beard](https://github.com/daniel-beard)
   [#256](https://github.com/realm/SwiftLint/issues/256)
 
-* Improve performance of `ColonRule`.  
+* Improve performance of `ColonRule` & `syntaxKindsByLine`.  
   [Norio Nomura](https://github.com/norio-nomura)
 
 * Add command to display rule description.
@@ -34,8 +34,7 @@
 
 ##### Enhancements
 
-* Improve performance of `syntaxKindsByLine`.  
-  [Norio Nomura](https://github.com/norio-nomura)
+* None.
 
 ##### Bug Fixes