This depends on https://github.com/jpsim/SourceKitten/pull/152
The duration of linting Carthage 0.11 is reduced from 17sec to 16sec
SwiftLint(32b325b) with SourceKitten(0f54e19):
```
swiftlint lint --config ~/.swiftlint-test.yml  17.48s user 1.12s system 83% cpu 22.182 total
```
SwiftLint(this) with SourceKitten(d7f9266):
```
swiftlint lint --config ~/.swiftlint-test.yml  16.05s user 1.05s system 85% cpu 20.051 total
```
On Instruments, `*BodyLengthRule`s are reduced from:
<img width="900" alt="screenshot 2016-01-30 13 42 20" src="https://cloud.githubusercontent.com/assets/33430/12693747/a944ddc4-c757-11e5-9f0b-1c947bffe8f2.png">
to:
<img width="901" alt="screenshot 2016-01-30 13 43 46" src="https://cloud.githubusercontent.com/assets/33430/12693748/b2dfbfb6-c757-11e5-9db2-45c4f3e30ae3.png">