@norio-nomura 2016-01-30
Oops, this and  #453 need bump SourceKitten for tests.