@norio-nomura 2016-01-30
I don't know how update SourceKitten only in `Cartfile.resolved` without pinning in `Cartfile`.
Is it need editing manually?