@jpsim 2016-01-30
It's ok to pin in Cartfile: `github "jpsim/SourceKitten" "master"` in this case, until a SourceKitten release is made.