@norio-nomura 2016-01-30
I ran `carthage update --no-use-binaries --no-build` with `Cartfile`:
```
github "jpsim/SourceKitten" "master"
```
and got `Cartfile.resolved`:
```diff
diff --git a/Cartfile.resolved b/Cartfile.resolved
index 85c2e92..425c818 100644
--- a/Cartfile.resolved
+++ b/Cartfile.resolved
@@ -1,6 +1,6 @@
-github "antitypical/Result" "1.0.1"
-github "drmohundro/SWXMLHash" "2.0.6"
+github "antitypical/Result" "1.0.2"
+github "drmohundro/SWXMLHash" "2.1.0"
 github "behrang/YamlSwift" "1.3.1"
 github "jspahrsummers/xcconfigs" "0.8.1"
 github "Carthage/Commandant" "0.8.2"
-github "jpsim/SourceKitten" "0.8.0"
+github "jpsim/SourceKitten" "4db2577f68d3f132509f34a72cc6dcfeda78ac85"
```
SourceKitten also does not use pinning in `Cartfile` and `Cartfile.private`.
So, it seems another modules also updated.