@jpsim 2016-01-30
Ah, right. Well it doesn't really matter, you can just edit the submodule like you did.