@norio-nomura 2016-01-30
OK, I will push with updated submodules on next time if it will depends on them, and won't think about `Cartfile*`. :smiley_cat: