@jpsim 2016-01-31
I've rebased this and will merge once CI passes. Thanks @norio-nomura!